Update 1% section
Because the total numbers have been updated, the previously calculated figure for donations from the top 1% of donors is likely no longer exactly accurate. This commit removes the exact figure that was listed and leaves the '41% of all donations' line on the assumption that this is still generally true. To be checked.